executor_a round4: numeric eclipse ArrayList @TestFactory consolidation
